+++ src/tests/obj-accumulator.mock.js
@@ -1,0 +1,8 @@
+module.exports = { method, getter, accumulator }
+
+function method() {}
+function getter() {}
+
+function accumulator() {
+  return { method, getter }
+}