+++ CHANGELOG.md
@@ -17,9 +17,10 @@
 
 ## 0.1.1 / 2019-05-08
 - update `README.md`
-- apply DI to defineAccumulator function
+- apply DI to `defineAccumulator` function
 - add restriction: cannot add object by name already present in storage
 - add requirement: throw error if try to get object by name not presented in storage
+- add mock for `accumulator` and rewrite unit tests for `defineAccumulator`
 
 ## 0.0.1 / 2019-05-07
 - Initial version v0.0.1